Bug 1440664 - Update gecko WR bindings for API changes in WR PR 2449. r=kats
MozReview-Commit-ID: AUTNkZ36p8b

UltraBlame original commit: c066c419e3789c50158bd83dd641b148c9fbb886

+++ gfx/webrender_bindings/WebRenderTypes.h
@@ -756,7 +756,7 @@ static inline wr::WrFilterOpType ToWrFilterOpType(uint32_t type) {
 
 
 struct WrClipId {
-  uint64_t id;
+  size_t id;
 
   bool operator==(const WrClipId& other) const {
     return id == other.id;
@@ -766,7 +766,7 @@ struct WrClipId {
 
 
 struct WrScrollId {
-  uint64_t id;
+  size_t id;
 
   bool operator==(const WrScrollId& other) const {
     return id == other.id;
@@ -780,7 +780,7 @@ struct WrScrollId {
 
 
 struct WrStickyId {
-  uint64_t id;
+  size_t id;
 
   bool operator==(const WrStickyId& other) const {
     return id == other.id;
